+++ src/org/usfirst/frc/team1014/robot/Robot.java
@@ -1,7 +1,7 @@
 package org.usfirst.frc.team1014.robot;
 
-import org.usfirst.frc.team1014.control.BadScheduler;
 import org.usfirst.frc.team1014.robot.commands.CommandBase;
+import org.usfirst.frc.team1014.robot.controls.BadScheduler;
 import org.usfirst.frc.team1014.utilities.SmartDashboard;
 
 import edu.wpi.first.wpilibj.IterativeRobot;
