+++ src/org/usfirst/frc/team1014/robot/controls/BadScheduler.java
@@ -1,6 +1,4 @@
-package org.usfirst.frc.team1014.control;
-
-import org.usfirst.frc.team1014.robot.controls.ControlsManager;
+package org.usfirst.frc.team1014.robot.controls;
 
 import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.command.Scheduler;
